requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic[email]==2.11.9
pydantic-settings==2.10.1
sqlalchemy==2.0.43
asyncpg==0.30.0
PyJWT==2.10.1
alembic==1.16.5
bcrypt==4.3.0
passlib==1.7.4
httpx==0.28.1
numpy==1.26.4
redis==6.4.0